Restructuration and dbConnection improvements
