+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/S3T3n1.main.iml" filepath="$PROJECT_DIR$/.idea/modules/S3T3n1.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/SQLsprint3.iml" filepath="$PROJECT_DIR$/.idea/SQLsprint3.iml" />
     </modules>
   </component>
