+++ .idea/modules/S3T3n1.main.iml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="AdditionalModuleElements">
+    <content url="file://$MODULE_DIR$/../../SQLsprint3/S3T3n1/src/main" dumb="true">
+      <sourceFolder url="file://$MODULE_DIR$/../../SQLsprint3/S3T3n1/src/main/resources" type="java-resource" />
+    </content>
+  </component>
+</module>
