+++ SQLsprint3/S3T3n1/src/main/java/org/example/model/repository/DatabaseConnection.java
@@ -1,41 +1,39 @@
 package org.example.model.repository;
 
-//import java.io.IOException; needed for inputstream
-//import java.io.InputStream;
+import java.io.FileInputStream;
+import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.util.Properties;
 
 public class DatabaseConnection {
-	
-    private static Connection connection;
-    private static Properties pr = new Properties(); 
-    private static String url = "jdbc:mysql://localhost:3306/SQLsprint3";
-    private static String username = "testUsername";
-    private static String password = "testPassword";
-    //private static InputStream stream;
-    
+    private static Properties properties = new Properties();
 
-    public static Connection getConnection() {
-    	
-        if (connection == null) {
-            try (Connection newConnect = DatabaseConnection.getConnection()) {
-            		//InputStream stream = getClass().getResourceAsStream("application.properties"))
-                connection = DriverManager.getConnection(url, username, password);
-            } catch (SQLException e) { // | IOException needed in catch block if inputstream is implemented
-                e.printStackTrace();
-            }
-        }
-        return connection;
+    static {
+        try (FileInputStream inputStream = new FileInputStream("application.properties")) {
+            properties.load(inputStream);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }//Made the properties loading static so it only loads once, when the DatabaseConnection class is called
     }
-    
+
+    public static Connection getConnection() throws SQLException {//we have to handle this exception in upper level wit try-with-resources
+        String url = properties.getProperty("db.url");
+        String username = properties.getProperty("db.username");
+        String password = properties.getProperty("db.password");
+
+        return DriverManager.getConnection(url, username, password);
+    }
+}
+
+
     /*
      * Singleton pattern needed for connection or is it already in use with the first null?
-     * 
-     * private DatabaseConnection() {} 
-     * 
-     * public static Connection getConnection() {	
+     *
+     * private DatabaseConnection() {}
+     *
+     * public static Connection getConnection() {
 		Connection result = connection;
 		if (result == null) {								//first null instance is to avoid multithreading issue, if instance exists already skips creation process below
 			synchronized (DatabaseConnection.class) {
@@ -47,7 +45,6 @@ public class DatabaseConnection {
 		}
 		return result;
 	}
-     * 
-     * 
+     *
+     *
      */
-}
