--- SQLsprint3/S3T3n1/src/main/java/org/example/model/repository/StockDAO.java
@@ -1,20 +1,0 @@
-package org.example.model.repository;
-
-import java.util.List;
-import org.example.model.domain.Product;
-
-public interface StockDAO {
-	
-	List<Product> findAll();
-	Product findById(int id);
-	List<Product> findByName(String name);
-	List<Product> findByPrice(double price);
-	List<Product> findByType(String type);
-	List<Product> findByHeight(float height);
-	List<Product> findByColor(String color);
-	List<Product> findByMaterial(boolean material);
-	boolean insertProduct(Product p);
-	boolean updateProduct(Product p);
-	boolean deleteProduct(int id);
-
-}
