+++ SQLsprint3/S3T3n1/src/main/java/org/example/model/service/Application.java
@@ -3,6 +3,7 @@ package org.example.model.service;
 import java.util.Scanner;
 
 import org.example.model.domain.*;
+import org.example.model.domain.entity.Product;
 
 public class Application {
