+++ SQLsprint3/S3T3n1/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+db.url=jdbc:mysql://localhost:3306/SQLsprint3
+db.username=testUsername
+db.password=testPassword
